admin-backend/pyproject.toml
[tool.flake8]
exclude= [".eggs", "./rag-core-library/*", "./src/admin_backend/models/*", "./src/admin_backend/rag_backend_client/*", "./src/admin_backend/document_extractor_client/*", ".git", ".hg", ".mypy_cache", ".tox", ".venv", ".devcontainer", "venv", "_build", "buck-out", "build", "dist"]
statistics = true
show-source = false
max-complexity = 8
max-annotations-complexity = 3
docstring-convent = 'numpy'
max-line-length = 120
ignore = ["E203", "W503", "E704"]
inline-quotes = '"'
docstring-quotes = '"""'
multiline-quotes = '"""'
dictionaries = ["en_US", "python", "technical", "pandas"]
ban-relative-imports = true
per-file-ignores = """./src/admin_backend/impl/prompt_templates/summarize_prompt.py: E501"""

[tool.black]
line-length = 120
exclude = """
/(
      .eggs
    | .git
    | .hg
    | .mypy_cache
    | .nox
    | .pants.d
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
    | node_modules
    | venv
)/
"""

[tool.isort]
profile = "black"
skip = ['.eggs', '.git', '.hg', '.mypy_cache', '.nox', '.pants.d', '.tox', '.venv', '_build', 'buck-out', 'build', 'dist', 'node_modules', 'venv']
skip_gitignore = true

[tool.pylint]
max-line-length = 120

[tool.poetry]
name = "admin_backend"
version = "0.0.1"
description = "The admin backend is responsible for the document management. This includes deletion, upload and getting particular documents or document lists."
authors = ["mmmake GmbH"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"
pytest = "^8.2.2"
flake8 = "^7.1.0"
flake8-black = "^0.3.6"
flake8-pyproject = "^1.2.3"
coverage = "^7.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
rag-core-lib = { path = "../rag-core-library/rag-core-lib", develop = true}
python = "^3.11"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
dependency-injector = "^4.41.0"
langchain-community = "^0.2.4"
pydantic-settings = "^2.3.3"
oauthlib = "^3.2.2"
requests-oauthlib = "^2.0.0"
python-dateutil = "^2.9.0.post0"
tenacity = "8.3.0"
boto3 = "^1.34.127"
tqdm = "^4.66.4"
langfuse = "^2.39.1"
